backend/requirements.txt
Django==6.0.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
psycopg[binary]==3.2.13
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
django-allauth==0.63.3
PyJWT==2.10.1
cryptography==42.0.0
django-filter==24.3
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.1
django-environ==0.11.2
requests==2.32.3
httpx==0.25.2

# Dev/Tests
pytest==8.2.1
pytest-django==4.8.0
model-bakery==1.18.1
flake8==6.1.0
black==24.1.1
isort==5.13.2

